requirements.txt
flask==3.0.3
selenium==4.15.2
webdriver-manager==4.0.1
openpyxl==3.1.2
python-docx==1.1.0
opencv-python==4.10.0.84
moviepy==1.0.3
google-generativeai==0.7.2
mss==9.0.1
pynput==1.7.6
tqdm==4.66.5